lock_free_hashtable::atomic_value

Trait AtomicValue

pub trait AtomicValue {
    type Raw: Copy;
    type Ref<'a>: Copy
       where Self: 'a;

    // Required methods
    fn null() -> Self::Raw;
    fn is_null(this: Self::Raw) -> bool;
    fn into_raw(this: Self) -> Self::Raw;
    unsafe fn from_raw(raw: Self::Raw) -> Self;
    unsafe fn deref<'a>(raw: Self::Raw) -> Self::Ref<'a>;
}

Generalized non-null pointer.

Anything which can be stored in LockFreeRawTable.

Required Associated Types

type Raw: Copy

The value stored in the table.

If the value is larger than max support atomic width on the platform, atomic operations will be spinlocked.

type Ref<'a>: Copy where Self: 'a

Dereferenced value.

Required Methods

fn null() -> Self::Raw

Null value. This is stored in LockFreeRawTable when the entry is missing.

fn is_null(this: Self::Raw) -> bool

Check if the value is null.

fn into_raw(this: Self) -> Self::Raw

Obtain the raw pointer. It must be non-null.

unsafe fn from_raw(raw: Self::Raw) -> Self

Construct the value from the raw pointer.

unsafe fn deref<'a>(raw: Self::Raw) -> Self::Ref<'a>

Dereference the raw pointer.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.